% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{AudioResponse}
\alias{AudioResponse}
\title{AudioResponse Class}
\description{
An R6Class object representing a response returned by the OpenAI Audio API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AudioResponse-new}{\code{AudioResponse$new()}}
\item \href{#method-AudioResponse-extract}{\code{AudioResponse$extract()}}
\item \href{#method-AudioResponse-print}{\code{AudioResponse$print()}}
\item \href{#method-AudioResponse-clone}{\code{AudioResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-AudioResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioResponse$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioResponse-extract"></a>}}
\if{latex}{\out{\hypertarget{method-AudioResponse-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioResponse$extract(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-AudioResponse-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AudioResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AudioResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AudioResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
