% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{ChatGPT_job}
\alias{ChatGPT_job}
\title{Run ChatGPT job}
\usage{
ChatGPT_job(
  viewer = "rstudioapi::viewer",
  api_url = NULL,
  api_key = NULL,
  organization = NULL
)
}
\arguments{
\item{viewer}{A viewer pane to use for displaying the gadget (defaults to paneViewer(minHeight = 300)).}

\item{api_url}{The URL of the OpenAI API.}

\item{api_key}{The API key to access the OpenAI API.}

\item{organization}{The ID of the organization used to access the OpenAI API.}
}
\description{
The function runs a ChatGPT job by calling the openapi::ChatGPT_gadget function with the specified parameters defined in the arguments. It also creates an R script file and writes the ChatGPT function code into the file. The R script file is then run as a job through the jobRunScript() function, which creates a background process that listens for user inputs and generates responses using OpenAI API.
}
\examples{
\dontrun{
# Set up OpenAI API endpoint URL and API key
api_url <- "https://api.openai.com"
api_key <- "my-secret-api-key"

# Start a ChatGPT job
ChatGPT_job(api_url = api_url, api_key = api_key)
}
}
\seealso{
\code{\link{ChatGPT_gadget}} function
}
