% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_requests.R
\name{check_response}
\alias{check_response}
\title{Check Response}
\usage{
check_response(resp, content_type = "application/json")
}
\arguments{
\item{resp}{HTTP response object returned by 'httr' package function.}

\item{content_type}{Expected content type of the response.}
}
\value{
Returns a string with value "ok" if response status code is not an error
code and the content type of the response matches the expected content type.
Otherwise it returns "notok".
}
\description{
Checks if the response was successful and matches the given content type.
}
\examples{
check_response(resp = httr::POST("http://httpbin.org/post", body = list(a = 1, b = "xyz")))
check_response(resp = httr::POST("http://httpbin.org/status/404"), content_type = "application/json")
}
