% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_comment}
\alias{code_comment}
\title{Add inline comments to the following R code to improve its readability.}
\usage{
code_comment(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing-class}} object
}
\description{
Add inline comments to the following R code to improve its readability.
}
