% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_create_test}
\alias{code_create_test}
\title{Create Test Units for R Code}
\usage{
code_create_test(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing}} object
}
\description{
Given a code snippet, generates test units using the testthat package.
}
