% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_document}
\alias{code_document}
\title{Generate complete roxygen2 documentation for R code.}
\usage{
code_document(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing}} object
}
\description{
Generate complete roxygen2 documentation for R code.
}
