% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_edit}
\alias{code_edit}
\title{Edit R code with a user prompt and system message}
\usage{
code_edit(code, prompt, instruction, explain = TRUE, simplify = TRUE, ...)
}
\arguments{
\item{code}{R code to be edited}

\item{prompt}{System prompt message, character vector}

\item{instruction}{User instruction message, character vector}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing-class}} object
}
\description{
This function returns a text editing object that allows the user to edit R code based on a prompt and message from the system. The user can choose to receive explanation or not. The edited code is then compared to the original code and the differences are returned.
}
