% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_explain}
\alias{code_explain}
\title{Explain R code}
\usage{
code_explain(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing}} object
}
\description{
Review the following R code and provide an explanation of how it works.
}
