% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_improve}
\alias{code_improve}
\title{Improve R code efficiency}
\usage{
code_improve(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing-class}} object
}
\description{
This function takes R code as input and edits it to improve its efficiency.
}
