% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{code_refactor}
\alias{code_refactor}
\title{Refactor R code for efficiency and readability}
\usage{
code_refactor(
  code,
  additional_instructions = NULL,
  explain = TRUE,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{code}{R code to be edited}

\item{explain}{Logical, indicating whether to include explanation messages}

\item{...}{Additional arguments to pass to \code{\link{create_chat_completion}}}
}
\value{
A \code{\link{TextEditing-class}} object
}
\description{
This function takes in R code and refactors it to improve its efficiency and readability.
}
