% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_edit.R
\name{compare_text}
\alias{compare_text}
\title{Compare Two Texts}
\usage{
compare_text(text1, text2)
}
\arguments{
\item{text1, text2}{Character vectors of text to be compared.}
}
\value{
A difference report between the two texts.
}
\description{
Compares two texts and returns a difference report.
}
\examples{
text1 <- "Lorem ipsum dolor sit amet."
text2 <- "Lorem ipsum sit amet."
compare_text(text1, text2)
}
