% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CompletionResponse}
\alias{CompletionResponse}
\title{CompletionResponse Class}
\description{
An R6Class object representing a response returned by the OpenAI Completion API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompletionResponse-new}{\code{CompletionResponse$new()}}
\item \href{#method-CompletionResponse-extract}{\code{CompletionResponse$extract()}}
\item \href{#method-CompletionResponse-print}{\code{CompletionResponse$print()}}
\item \href{#method-CompletionResponse-clone}{\code{CompletionResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionResponse$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionResponse-extract"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionResponse-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionResponse$extract(field = "id")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionResponse-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompletionResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompletionResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompletionResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
