% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{create_chat_completion}
\alias{create_chat_completion}
\title{Create chat completions using the OpenAI API.}
\usage{
create_chat_completion(
  endpoint = "v1/chat/completions",
  model = "gpt-3.5-turbo",
  messages,
  temperature = 1,
  top_p = 1,
  n = 1,
  stream = TRUE,
  stop = NULL,
  max_tokens = NULL,
  presence_penalty = 0,
  frequency_penalty = 0,
  logit_bias = NULL,
  user = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{Endpoint for the OpenAI API. Default is "v1/chat/completions".}

\item{model}{ID of the GPT-3 model to be used. Default is "gpt-3.5-turbo".}

\item{messages}{A string or a list of strings representing the conversation history.}

\item{temperature}{Controls the degree of randomness of the response. Default is 1.}

\item{top_p}{Controls the number of tokens to consider for each response. Default is 1.}

\item{n}{Number of completions to generate. Default is 1.}

\item{stream}{Logical, whether to use websocket or not for communicating with OpenAI API. Default is FALSE.}

\item{stop}{One or more strings where each string represents a sequence at which the response should end.}

\item{max_tokens}{Controls the maximum number of tokens to generate in the response. Default is NULL (unrestricted).}

\item{presence_penalty}{Controls the degree to which the response is influenced by the presence of certain words in the conversation history. Default is 0.}

\item{frequency_penalty}{Controls the degree to which the response is influenced by the frequency of certain words in the conversation history. Default is 0.}

\item{logit_bias}{Offset the logits that OpenAI calculates for each token. Default is NULL.}

\item{user}{A string identifying the user whose conversation history is being used. Default is NULL.}

\item{max_tries}{The maximum number of attempts to make for a single API call. Default is 1.}

\item{timeout}{The maximum time in seconds to wait for a response from the OpenAI API. Default is 300.}

\item{...}{Additional arguments to be passed to 'making_requests()' function.}

\item{api_url}{URL to the OpenAI API. Can be set to a custom endpoint. Default is NULL.}

\item{api_key}{API key for authentication with the OpenAI API. Default is NULL.}

\item{organization}{Organization ID for authentication with the OpenAI API. Default is NULL.}
}
\value{
Response object of class 'CompletionResponse'.
}
\description{
This function allows users to generate a response to a given conversation history using OpenAI's API.
}
