% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings.R
\name{create_embeddings}
\alias{create_embeddings}
\title{Create embeddings using OpenAI API}
\usage{
create_embeddings(
  endpoint = "v1/embeddings",
  model = "text-embedding-ada-002",
  input,
  user = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{The API endpoint. Default: "v1/embeddings".}

\item{model}{The model used for generating embeddings. Default: "text-embedding-ada-002".}

\item{input}{The text input for which embeddings need to be generated.}

\item{user}{Optional user identifier.}

\item{max_tries}{Maximum number of attempts to make the API request in case of failure. Default: 1.}

\item{timeout}{Timeout in seconds for the API request. Default: 300.}

\item{api_url}{The OpenAI API URL. Default: NULL.}

\item{api_key}{Optional API key for authentication.}

\item{organization}{Optional organization identifier.}
}
\value{
An object of class "EmbeddingsResponse".
}
\description{
This function uses the OpenAI API to create embeddings for the given input text.
}
