% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image}
\alias{create_image}
\title{Create image using OpenAI's DALL-E API}
\usage{
create_image(
  endpoint = "images/generations",
  prompt,
  n = 1,
  size = c("1024x1024", "512x512", "256x256"),
  response_format = c("url", "b64_json"),
  user = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{the DALL-E API endpoint to use. Default is "images/generations".}

\item{prompt}{the prompt to send to the API. Required.}

\item{n}{the number of images to generate. Default is 1.}

\item{size}{the size of the image(s) to generate. Allowed values are "1024x1024", "512x512", and "256x256". Default is "1024x1024".}

\item{response_format}{the response format. Allowed values are "url" and "b64_json". Default is "url".}

\item{user}{the user to whom the API key belongs. Optional.}

\item{max_tries}{maximum number of retries in case of a network error. Default is 1.}

\item{timeout}{time in seconds to wait for the API response. Default is 300 seconds.}

\item{api_base}{the url of the DALL-E API server. Optional.}

\item{api_key}{the API key to use for the API. Required.}

\item{organization}{the organization to which the API key belongs. Optional.}
}
\value{
An object of class "ImagesResponse".
}
\description{
This function creates images using OpenAI's DALL-E API by sending the prompt to the API and specifying the size, number of images to generate and response format.
}
