% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image_edit}
\alias{create_image_edit}
\title{Create Image Edit}
\usage{
create_image_edit(
  endpoint = "v1/images/edits",
  image,
  mask = NULL,
  prompt,
  n = 1,
  size = c("1024x1024", "512x512", "256x256"),
  response_format = c("url", "b64_json"),
  user = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{endpoint for API request (default: "v1/images/edits")}

\item{image}{raw image data or a file path to the image.}

\item{mask}{raw image data or path to mask image.}

\item{prompt}{prompt describing edits to make to image.}

\item{n}{integer, number of images to generate (default: 1)}

\item{size}{dimensions of image, must be one of c("1024x1024", "512x512", "256x256") (default: "1024x1024")}

\item{response_format}{format for the response, must be one of c("url", "b64_json") (default: "url")}

\item{user}{user the request is associated with, if any.}

\item{max_tries}{Maximum number of tries to make request.}

\item{timeout}{Maximum time to wait for response.}

\item{api_url}{The API URL to query. The default value is "https://api.openai.com".}

\item{api_key}{Authentication key for API.}

\item{organization}{The name of the organization to use with this request. \code{api_key} must have permission to use this organization.}
}
\value{
An object of class "ImagesResponse".
}
\description{
Allows you to create an edited image by providing a raw image and a prompt that describes the edits you want to make. You can optionally include a mask image to control where the edits apply.
}
