% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{create_image_variation}
\alias{create_image_variation}
\title{Create Image Variation}
\usage{
create_image_variation(
  endpoint = "v1/images/variations ",
  image,
  n = 1,
  size = c("1024x1024", "512x512", "256x256"),
  response_format = c("url", "b64_json"),
  user = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{The API endpoint to use (default is \code{"v1/images/variations"}).}

\item{image}{The path to the image file.}

\item{n}{The number of variations to create (default is 1).}

\item{size}{A character string giving the desired size of the image variation (default is \code{"1024x1024"}). Possible values are \code{"1024x1024"}, \code{"512x512"}, and \code{"256x256"}.}

\item{response_format}{The format in which the image variation is returned (default is \code{"url"}). Possible values are \code{"url"} (returns the URL to the image variation) and \code{"b64_json"} (returns the base64 encoded JSON object to the image variation).}

\item{user}{The user making the request (default is NULL).}

\item{max_tries}{The maximum number of times to try the request (default is 1).}

\item{timeout}{The timeout for the request in seconds (default is 300).}

\item{api_url}{The URL of the API (default is NULL).}

\item{api_key}{The API key to use for authentication (default is NULL).}

\item{organization}{The organization making the request (default is NULL).}
}
\value{
An object of class "ImagesResponse".
}
\description{
The function creates variations of an image based on size parameter and returns it in URL or base64 encoded JSON format.
}
