% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_transcription}
\alias{create_transcription}
\title{Create Transcription}
\usage{
create_transcription(
  endpoint = "v1/audio/transcriptions",
  file,
  model = "whisper-1",
  prompt = NULL,
  response_format = c("json", "text", "srt", "verbose_json", "vtt"),
  temperature = 0,
  language = NULL,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{an endpoint url (default "v1/audio/transcriptions").}

\item{file}{name and path of the audio file.}

\item{model}{name of the language model to use (default "whisper-1").}

\item{prompt}{a string prompt which will be prepended to the file transcription.}

\item{response_format}{a character string indicating the type of response format (default "json").}

\item{temperature}{a parameter controlling the creativity of the model. Set between 0 and 1 (default 0).}

\item{language}{the language code of the audio if known. Currently this parameter
only supports English ("en") or Mandarin ("zh") (default NULL)}

\item{max_tries}{the maximum number of times to try the request if there are connection errors (default 1).}

\item{timeout}{maximum time to wait for a reponse (default 300).}

\item{api_url}{base URL for the OpenAI API (default NULL).}

\item{api_key}{An API Key is required to use this tool.}

\item{organization}{Optional organization slug required for use with OpenAI API.}
}
\value{
An object of class "AudioResponse".
}
\description{
This function creates transcription of an audio file using OpenAI language models
}
