% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{create_translation}
\alias{create_translation}
\title{Create Translation}
\usage{
create_translation(
  endpoint = "v1/audio/translations",
  file,
  model = "whisper-1",
  prompt = NULL,
  response_format = c("json", "text", "srt", "verbose_json", "vtt"),
  temperature = 0,
  api_url = NULL,
  api_key = NULL,
  organization = NULL,
  max_tries = 1,
  timeout = 300
)
}
\arguments{
\item{endpoint}{A string indicating the API endpoint for making requests (default is "v1/audio/translations").}

\item{file}{A character string indicating the file path of the audio file to be translated.}

\item{model}{A character string indicating the model name (default is "whisper-1").}

\item{prompt}{A string to use as input prompt.}

\item{response_format}{A character string indicating the response format (accepts "json", "text", "srt", "verbose_json", or "vtt").}

\item{temperature}{A numeric value indicating the degree of randomness in the model's output (default is 0).}

\item{api_url}{A string indicating the base URL of the OpenAI API (default is NULL).}

\item{api_key}{A string indicating the authentication key to use for requests (default is NULL).}

\item{organization}{A string indicating the organization ID (default is NULL).}

\item{max_tries}{An integer value indicating the maximum number of attempts to make the API call (default is 1).}

\item{timeout}{A numeric value indicating the timeout limit in seconds for the API call (default is 300).}
}
\value{
An object of class "AudioResponse".
}
\description{
This function translates an audio file to text using OpenAI API.
}
