% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{div_create}
\alias{div_create}
\title{Create a list of HTML divs for a chat application}
\usage{
div_create(messages, openai_logo, user_logo)
}
\arguments{
\item{messages}{A list of messages containing content, role, and time.}

\item{openai_logo}{OpenAI logo content.}

\item{user_logo}{User logo content.}
}
\value{
A list of HTML divs.
}
\description{
This function takes in a list of messages and creates HTML divs
for each message in a chat application. It also includes the
HTML div for displaying ChatGPT output.
}
