% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download}
\alias{download}
\title{Download a file from a URL}
\usage{
download(
  url,
  destfile,
  methods = c("auto", "wget", "libcurl", "curl", "wininet", "internal"),
  quiet = TRUE,
  ...,
  max_tries = 1
)
}
\arguments{
\item{url}{a character string specifying the URL to download the file from}

\item{destfile}{a character string specifying the name of the file in which to save the download.}

\item{methods}{A character string or vector specifying the methods to be used for the transfer. Default is "auto".}

\item{quiet}{a logical indicating whether to run quietly, default TRUE}

\item{...}{optional additional arguments passed to download.file.}

\item{max_tries}{an integer indicating the maximum number of retries to attempt for downloading the file. Default is 1.}
}
\value{
`NULL`. The function invisibly returns NULL.
}
\description{
Downloads a file from a given URL and saves it to a specified destination file.
}
\examples{
\dontrun{
# Download MP3 file:
download("http://www.example.com/audio.mp3", "audio.mp3")
# Download CSV file:
download("http://www.example.com/mydata.csv", "mydata.csv")
}
}
