% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{EmbeddingsResponse}
\alias{EmbeddingsResponse}
\title{EmbeddingsResponse Class}
\description{
An R6Class object representing a response returned by the OpenAI Embeddings API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EmbeddingsResponse-new}{\code{EmbeddingsResponse$new()}}
\item \href{#method-EmbeddingsResponse-extract}{\code{EmbeddingsResponse$extract()}}
\item \href{#method-EmbeddingsResponse-print}{\code{EmbeddingsResponse$print()}}
\item \href{#method-EmbeddingsResponse-clone}{\code{EmbeddingsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsResponse$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsResponse-extract"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsResponse-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsResponse$extract(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsResponse-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EmbeddingsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EmbeddingsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EmbeddingsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
