% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_image}
\alias{fetch_image}
\title{Downloads an image from a given URL and displays it if desired.}
\usage{
fetch_image(url, destfile = NULL, plot = TRUE)
}
\arguments{
\item{url}{the URL of the image to download.}

\item{destfile}{(optional) path to the destination file where the downloaded image should be saved. If not specified, a temporary file with a file extension that matches the URL will be created, and automatically deleted when the function finishes.}

\item{plot}{(optional) whether to display the downloaded image or not. Default is TRUE.}
}
\value{
An image object of class "grob".
}
\description{
Downloads an image from a given URL and displays it if desired.
}
\examples{
fetch_image("https://www.r-project.org/logo/Rlogo.png")

}
