% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_prompts}
\alias{fetch_prompts}
\title{Fetch Chat Prompts}
\usage{
fetch_prompts(language = c("en", "zh"))
}
\arguments{
\item{language}{A character string specifying the language of the chat prompts you want to fetch (default "en").}
}
\value{
A data frame containing the chat prompts in the specified language.
}
\description{
This function fetches chat prompts from Github for various languages.
}
\examples{
prompts <- fetch_prompts(language = "en")
head(prompts)
}
