% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ImagesResponse}
\alias{ImagesResponse}
\title{ImagesResponse Class}
\description{
An R6Class object representing a response returned by the OpenAI Images API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagesResponse-new}{\code{ImagesResponse$new()}}
\item \href{#method-ImagesResponse-extract}{\code{ImagesResponse$extract()}}
\item \href{#method-ImagesResponse-print}{\code{ImagesResponse$print()}}
\item \href{#method-ImagesResponse-clone}{\code{ImagesResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesResponse$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesResponse-extract"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesResponse-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesResponse$extract(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesResponse-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagesResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagesResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagesResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
