% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{list_models}
\alias{list_models}
\title{List Models}
\usage{
list_models(endpoint = "models", max_tries = 1, timeout = 300, ...)
}
\arguments{
\item{endpoint}{The endpoint for the API request. Default is "models"}

\item{max_tries}{Maximum number of attempts for making the request. Default is 1.}

\item{timeout}{Timeout limit for the API request in seconds. Default is 300.}

\item{api_base}{The base URL of the OpenAI API. Default is NULL.}

\item{api_key}{Your API key for authentication. Default is NULL.}

\item{organization}{Organization ID for using resources from a specific organization. Default is NULL.}
}
\value{
Response object of class 'ModelsResponse'.
}
\description{
Retrieves a list of available models from OpenAI API.
}
