% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_requests.R
\name{making_requests}
\alias{making_requests}
\title{Make HTTP requests to the API endpoint}
\usage{
making_requests(
  method = c("POST", "GET", "DELETE"),
  endpoint = "chat/completions",
  data = NULL,
  encode = "json",
  stream = FALSE,
  stream_type = c("completion", "chat_completion"),
  stream_file = NULL,
  post_type = "application/json",
  response_type = "application/json",
  api_base = NULL,
  api_key = NULL,
  api_type = NULL,
  api_version = NULL,
  organization = NULL,
  azure_deployment = NULL,
  max_tries = 1,
  timeout = 300,
  debug = FALSE
)
}
\arguments{
\item{method}{Character string, defaults to "POST". The HTTP Method (POST, GET, DELETE) to use in making the request.}

\item{endpoint}{A character string to indicate the target URL endpoint.}

\item{data}{A named list or an object that can be serialized to JSON containing the input data to send/request to/from the endpoint.}

\item{encode}{A character string, defaults to "json". The encoding to use in packaging the text to be sent to the server.}

\item{stream}{A logical flag indicating whether to stream the content or not. It defaults to \code{FALSE}.}

\item{stream_type}{A character string, defaults to "completion". Indicates the type of data to be sent when streaming, either \code{"completion"} or \code{"chat_completion"}.}

\item{stream_file}{A character string indicating the name and path of a file to write the response data to.}

\item{post_type}{A character string, defaults to "application/json". Indicates the MIME type to be used as the content type of the request stream.}

\item{response_type}{A character string, defaults to "application/json". Indicates the MIME type to be used as the content type of the server response.}

\item{api_base}{A character string indicating the URL for the server endpoint.}

\item{api_key}{A character string indicating the API key to use for authentication.}

\item{organization}{A character string indicating the organization ID to use for authentication.}

\item{max_tries}{An integer, defaults to 1. The maximum number of times to try making the request before outputting an error.}

\item{timeout}{An integer, defaults to 300 seconds. The maximum time in seconds allowed for making the request.}

\item{debug}{A logical flag indicating whether to print debugging output or not. Defaults to \code{FALSE}.}

\item{key_nm}{A character string indicating the name of the authentication key in the headers.}

\item{organization_nm}{A character string indicating the name of the organization ID key in the headers.}
}
\value{
A response object returned by the server endpoint.
}
\description{
This function helps in making HTTP requests to the OpenAI API endpoint. It sends the encoded data provided to it and sets up headers including the authentication header to use in making the authentication. It tries to make the request multiple times based on the supplied values for the \code{max_tries} argument. It also provides functionality for streaming data and writing it to file.
}
