% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mask_image}
\alias{mask_image}
\title{Mask an image based on RGBA ranges}
\usage{
mask_image(
  image,
  height_range = NULL,
  width_range = NULL,
  R_range = NULL,
  G_range = NULL,
  B_range = NULL,
  destfile = NULL,
  plot = TRUE
)
}
\arguments{
\item{image}{The path to the image file}

\item{height_range}{The range of pixel heights to mask out (optional)}

\item{width_range}{The range of pixel widths to mask out (optional)}

\item{R_range}{The range of red pixel values to mask out (optional)}

\item{G_range}{The range of green pixel values to mask out (optional)}

\item{B_range}{The range of blue pixel values to mask out (optional)}

\item{destfile}{The path and filename to save the masked image (optional)}

\item{plot}{Logical value indicating whether to plot the masked image (default = TRUE)}
}
\value{
A rasterGrob object containing the masked image
}
\description{
Mask an image based on RGBA ranges
}
\examples{
\dontrun{
# Mask out a range of pixel heights and widths
mask_image("example.png",
  height_range = c(100, 200), width_range = c(50, 150)
)
# Mask out a range of RGB values
mask_image("example.png",
  R_range = c(0.2, 0.4), G_range = c(0, 0.3), B_range = c(0.1, 0.5)
)
# Save the masked image
mask_image("example.png",
  height_range = c(100, 200), width_range = c(50, 150),
  destfile = "masked_image.png"
)
}

}
