% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{ModelsResponse}
\alias{ModelsResponse}
\title{ModelsResponse Class}
\description{
An R6Class object representing a response returned by the OpenAI Models API
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelsResponse-new}{\code{ModelsResponse$new()}}
\item \href{#method-ModelsResponse-extract}{\code{ModelsResponse$extract()}}
\item \href{#method-ModelsResponse-print}{\code{ModelsResponse$print()}}
\item \href{#method-ModelsResponse-clone}{\code{ModelsResponse$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsResponse-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsResponse-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsResponse$new(response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsResponse-extract"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsResponse-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsResponse$extract(field = "id", filter = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsResponse-print"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsResponse-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsResponse$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelsResponse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelsResponse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelsResponse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
