% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_requests.R
\name{parse_response}
\alias{parse_response}
\title{Parse Response}
\usage{
parse_response(resp, class = NULL)
}
\arguments{
\item{resp}{Response object to parse.}

\item{class}{Class to use for parsing.}
}
\value{
Object parsed with the specified class if class is specified and response is JSON, otherwise the contents of the response object.
}
\description{
This function parses a response using the specified class.
}
\examples{
\dontrun{
# parse response with specified class
parse_response(resp, MyResponseClass)
# parse response without class
parse_response(resp)
}
}
