% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{retrieve_file}
\alias{retrieve_file}
\title{Retrieve a file based on the file ID}
\usage{
retrieve_file(
  endpoint = "v1/files",
  file_id,
  api_url = NULL,
  api_key = NULL,
  organization = NULL,
  max_tries = 1,
  timeout = 300
)
}
\arguments{
\item{endpoint}{The API endpoint. Default is "v1/files".}

\item{file_id}{The ID of the file to be deleted.}

\item{api_url}{The OpenAI API URL. Default is "NULL.}

\item{api_key}{The OpenAI API key. By default is NULL.}

\item{organization}{The OpenAI organization. By default is NULL.}

\item{max_tries}{The maximum number of times to try the API request. Default is 1.}

\item{timeout}{The timeout in seconds for API requests. Default is 300 seconds.}
}
\description{
This function retrieves a specific file from the OpenAI API based on the file ID.
}
