% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{retrieve_file_content}
\alias{retrieve_file_content}
\title{Retrieve File Content}
\usage{
retrieve_file_content(
  endpoint = "files",
  file_id,
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{The API endpoint. Default is "files".}

\item{file_id}{The ID of the file to be deleted.}

\item{max_tries}{The maximum number of times to try the API request. Default is 1.}

\item{timeout}{The timeout in seconds for API requests. Default is 300 seconds.}
}
\description{
This function retrieves the content of a file from the OpenAI API.
}
