% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{TextEditing}
\alias{TextEditing}
\title{TextEditing
An R6Class for text editing}
\description{
TextEditing
An R6Class for text editing

TextEditing
An R6Class for text editing
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TextEditing-new}{\code{TextEditing$new()}}
\item \href{#method-TextEditing-extract}{\code{TextEditing$extract()}}
\item \href{#method-TextEditing-print}{\code{TextEditing$print()}}
\item \href{#method-TextEditing-clone}{\code{TextEditing$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextEditing-new"></a>}}
\if{latex}{\out{\hypertarget{method-TextEditing-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextEditing$new(input = NULL, response)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextEditing-extract"></a>}}
\if{latex}{\out{\hypertarget{method-TextEditing-extract}{}}}
\subsection{Method \code{extract()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextEditing$extract(field = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextEditing-print"></a>}}
\if{latex}{\out{\hypertarget{method-TextEditing-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextEditing$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TextEditing-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TextEditing-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextEditing$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
