% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_text}
\alias{truncate_text}
\title{Truncate Text to a Maximum Length}
\usage{
truncate_text(text, max_length = 50)
}
\arguments{
\item{text}{Character vector of text to truncate.}

\item{max_length}{Maximum desired length of the truncated text. Defaults to 50.}
}
\value{
A character vector of the truncated text.
}
\description{
Truncates text to a maximum length and adds an indicator of truncation.
}
\examples{
truncate_text(c("This is a long piece of text", "Shorter text"), max_length = 10)

}
