% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{try_get}
\alias{try_get}
\title{Try evaluating an expression with a specified number of tries.}
\usage{
try_get(
  expr,
  max_tries = 1,
  sleep = 1,
  error_message = "",
  retry_message = "Retrying..."
)
}
\arguments{
\item{expr}{the expression to evaluate}

\item{max_tries}{the maximum number of tries before an error is returned}

\item{sleep}{the number of seconds to wait before retrying}

\item{error_message}{the message to print when an error is caught}

\item{retry_message}{the message to print when a retry is attempted}
}
\value{
Returns the output of the evaluated expression or an error message.
}
\description{
Try evaluating an expression with a specified number of tries.
}
