% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Upload File to OpenAI API}
\usage{
upload_file(
  endpoint = "files",
  file,
  purpose = "fine-tune",
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{The API endpoint to use (default: "files")}

\item{file}{The file path to upload}

\item{purpose}{The intended purpose for uploading the file (default: "fine-tune")}

\item{max_tries}{The maximum number of times to attempt the API call (default: 1)}

\item{timeout}{The timeout value for the API call in seconds (default: 300)}

\item{api_base}{The URL of the OpenAI API (default: NULL)}

\item{api_key}{The API key for accessing the OpenAI API (default: NULL)}

\item{organization}{The organization associated with the API key (default: NULL)}
}
\description{
Uploads a file to the OpenAI API for fine-tuning models, among other purposes.
}
