% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_setup}
\alias{api_setup}
\title{API Setup}
\usage{
api_setup(
  api_url = "https://api.openai.com",
  api_key = NULL,
  organization = NULL,
  key_nm = "Authorization",
  organization_nm = "OpenAI-Organization",
  endpoint = NULL,
  check_url = TRUE
)
}
\arguments{
\item{api_url}{The URL of the OpenAI API. Default is "https://api.openai.com".}

\item{api_key}{The API key for accessing the OpenAI API.}

\item{organization}{The organization name associated with the OpenAI API key.}

\item{check_url}{Logical to check the validity of the specified URL. Default is TRUE.}
}
\value{
The function returns nothing (invisible NULL).
}
\description{
This function sets up the API URL, API key, and organization for OpenAI's services.
}
\examples{
\dontrun{
## Official OpenAI API
api_url <- "https://api.openai.com"
api_key <- "Bearer sk-xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
api_setup(api_url = api_url, api_key = api_key)

## Azure OpenAI API
api_url <- "https://xxxxxx.openai.azure.com"
api_key <- "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
api_setup(api_url = api_url, api_key = api_key, key_nm = "api-key")
}
}
