% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{delete_file}
\alias{delete_file}
\title{Delete a file from OpenAI API}
\usage{
delete_file(
  endpoint = "files",
  file_id,
  purpose = "fine-tune",
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{The API endpoint. Default is "files".}

\item{file_id}{The ID of the file to be deleted.}

\item{purpose}{The purpose of the file. Default is "fine-tune".}

\item{max_tries}{The maximum number of times to try the API request. Default is 1.}

\item{timeout}{The timeout in seconds for API requests. Default is 300 seconds.}

\item{api_base}{The OpenAI API URL. Default is "NULL.}

\item{api_key}{The OpenAI API key. By default is NULL.}

\item{organization}{The OpenAI organization. By default is NULL.}
}
\description{
This function allows you to delete a file stored in OpenAI API.
}
