% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{list_models}
\alias{list_models}
\title{List Models}
\usage{
list_models(
  endpoint = "v1/models",
  api_url = NULL,
  api_key = NULL,
  organization = NULL,
  max_tries = 1,
  timeout = 300
)
}
\arguments{
\item{endpoint}{The endpoint for the API request. Default is "v1/models"}

\item{api_url}{The base URL of the OpenAI API. Default is NULL.}

\item{api_key}{Your API key for authentication. Default is NULL.}

\item{organization}{Organization ID for using resources from a specific organization. Default is NULL.}

\item{max_tries}{Maximum number of attempts for making the request. Default is 1.}

\item{timeout}{Timeout limit for the API request in seconds. Default is 300.}
}
\value{
Response object of class 'ModelsResponse'.
}
\description{
Retrieves a list of available models from OpenAI API.
}
