% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_requests.R
\name{making_requests}
\alias{making_requests}
\title{Making Requests Function}
\usage{
making_requests(
  method = c("POST", "GET", "DELETE"),
  endpoint = "v1/chat/completions",
  data = NULL,
  encode = "json",
  stream = FALSE,
  stream_type = c("completion", "chat_completion"),
  stream_file = NULL,
  post_type = "application/json",
  response_type = "application/json",
  api_url = NULL,
  api_key = NULL,
  organization = NULL,
  max_tries = 1,
  timeout = 300,
  debug = FALSE
)
}
\arguments{
\item{method}{A character string indicating the http method to be used (default is "POST").}

\item{endpoint}{A character string indicating the endpoint to which the request should be sent (default is "v1/chat/completions").}

\item{data}{A list object containing the data to be sent with the request (default is NULL).}

\item{encode}{A character string specifying the encoding type for the data to be sent (default is "json").}

\item{stream}{A logical value indicating whether the response should be streamed or not (default is FALSE).}

\item{stream_type}{A character string specifying the type of stream content to expect in the response (default is "completion").}

\item{stream_file}{A character string indicating the file path to save the streamed content to (default is NULL).}

\item{post_type}{A character string specifying the content type of the request body (default is "application/json").}

\item{response_type}{A character string specifying the content type of the response body (default is "application/json").}

\item{api_url}{A character string indicating the OpenAI API endpoint URL (default is NULL).}

\item{api_key}{A character string indicating the user's API key (default is NULL).}

\item{organization}{A character string indicating the name of the OpenAI organization being accessed (default is NULL).}

\item{max_tries}{An integer specifying the maximum number of tries allowed for a given request before an error is thrown (default is 1).}

\item{timeout}{An integer specifying the timeout window for the request in seconds (default is 300).}

\item{debug}{A logical value indicating whether debug mode should be initiated (default is FALSE).}
}
\value{
A response object containing the http status code and response content.
}
\description{
This function allows users to make requests to OpenAI's API endpoint using various http methods.
}
