% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{retrieve_model}
\alias{retrieve_model}
\title{Retrieve OpenAI Model}
\usage{
retrieve_model(
  endpoint = "models",
  model = "text-davinci-003",
  max_tries = 1,
  timeout = 300,
  ...
)
}
\arguments{
\item{endpoint}{a character string indicating the endpoint, default is "models"}

\item{model}{a character string indicating the model name, default is "text-davinci-003"}

\item{max_tries}{an integer value indicating maximum number of tries, default is 1}

\item{timeout}{a numeric value indicating request timeout in seconds, default is 300}

\item{api_base}{a character string indicating the OpenAI API URL, default is NULL}

\item{api_key}{a character string for the OpenAI API key, default is NULL}

\item{organization}{a character string indicating the API organization, default is NULL}
}
\value{
Response object of class 'ModelsResponse'.
}
\description{
Retrieves details about a specific OpenAI model.
}
